+++ src/media_tools/avilib.c
@@ -123,8 +123,8 @@ static void short2str(unsigned char *dst, s32 n)
 
 static u64 str2ullong(unsigned char *str)
 {
-	u64 r = (str[0] | (str[1]<<8) | (str[2]<<16) | (str[3]<<24));
-	u64 s = (str[4] | (str[5]<<8) | (str[6]<<16) | (str[7]<<24));
+	u64 r = ((u32)str[0] | ((u32)str[1]<<8) | ((u32)str[2]<<16) | ((u32)str[3]<<24));
+	u64 s = ((u32)str[4] | ((u32)str[5]<<8) | ((u32)str[6]<<16) | ((u32)str[7]<<24));
 #ifdef __GNUC__
 	return ((s<<32)&0xffffffff00000000ULL)|(r&0xffffffff);
 #else
@@ -134,11 +134,11 @@ static u64 str2ullong(unsigned char *str)
 
 static u32 str2ulong(unsigned char *str)
 {
-	return ( str[0] | (str[1]<<8) | (str[2]<<16) | (str[3]<<24) );
+	return ( (u32)str[0] | ((u32)str[1]<<8) | ((u32)str[2]<<16) | ((u32)str[3]<<24) );
 }
 static u32 str2ushort(unsigned char *str)
 {
-	return ( str[0] | (str[1]<<8) );
+	return ( (u32)str[0] | ((u32)str[1]<<8) );
 }
 
 // bit 31 denotes a keyframe
